Replace hardcoded localhost URL with env variable in hands.http
Hardcoded URLs were replaced with an environment variable in hands.http to allow switching between different environments e.g. dev and prod. The environment variable "host" is now used instead of the hardcoded "localhost" URL. This ensures more flexibility and makes it easier to deploy the application in different environments without the need to manually change the code.